requirements.txt
pandas
scikit-learn==1.2.2
matplotlib
numpy
gplearn
transformers>=4.40.0
torch>=2.1.0
accelerate
google-cloud-bigquery
google-auth
pyarrow
db-dtypes
hydra-core>=1.3